feat(packages/sui-js): change naming and context
