+++ packages/sui-js/src/ua-parser/index.js
@@ -1,5 +1,14 @@
 import bowser from 'bowser'
 
+export const LEGITIMATE_CRAWLER_USER_AGENTS = [
+  'googlebot',
+  'google-structured-data-testing-tool',
+  'bingbot',
+  'linkedinbot',
+  'mediapartners-google',
+  'debugbear'
+]
+
 export const stats = userAgent => {
   const ua = bowser.parse(userAgent)
   return {
@@ -9,4 +18,12 @@ export const stats = userAgent => {
     browserVersion: ua.browser.version,
     isTablet: ua.platform.type === bowser.PLATFORMS_MAP.tablet
   }
+}
+
+export const checkLegitimateCrawler = (userAgent, crawlerUserAgents) => {
+  const lowerCaseUserAgent = userAgent.toLowerCase()
+
+  const crawlerUserAgentsList = crawlerUserAgents || LEGITIMATE_CRAWLER_USER_AGENTS
+
+  return crawlerUserAgentsList.some(ua => lowerCaseUserAgent.includes(ua))
 }
